refactor(tui): extract text, debug log, and clock helpers
First step of splitting the 12k-line tui module. Moves the width-aware
string helpers, the P2PMUX_DEBUG_UI trace log, and the unix clock read
into their own files, byte for byte, and re-exports them from the module
root so every existing path keeps resolving.
